Moved useValidation to own folder
