+++ src/hooks/useValidation/useValidation.ts
@@ -7,13 +7,13 @@
 
 import { useCallback, useState, useMemo } from 'react';
 
-import { TFieldErrors, TValidator, isDefaultValidator, TAsyncValidator, isIFieldErrorObject } from '../validators';
-import { parseValidationError } from '../utils';
+import { TFieldErrors, TValidator, isDefaultValidator, TAsyncValidator, isIFieldErrorObject } from '../../validators';
+import { parseValidationError } from '../../utils';
 
-import { useFormContext } from './useFormContext';
-import { TBasicFieldValue } from './useField';
-import { useIsUnmounted } from './useIsUnmounted';
-import { useTimeout } from './useTimeout';
+import { useFormContext } from '../useFormContext';
+import { TBasicFieldValue } from '../useField';
+import { useIsUnmounted } from '../useIsUnmounted';
+import { useTimeout } from '../useTimeout';
 
 export interface IBasicValidationState {
   /**
